Fix difference from working tree to master
