+++ .gitignore
@@ -59,3 +59,4 @@ yarn-error.log*
 /dump.rdb
 /supervisord.pid
 log
+\.vscode/
