+++ bsx3/ui/datacollection/dataC.css
@@ -68,7 +68,7 @@
 
 .horizontal-tabs .sv-tab-panel {
     background: #fff;
-    height: 1000px;
+    height: 900px;
     padding-top: 10px;
     padding: 10px
 }
